refactor create and edit function and view

+++ src/Base/AdminBundle/Service/FileUploader.php
@@ -17,9 +17,24 @@ class FileUploader
 
     /**
      * FileUploader constructor.
-     * @param $targetDir
      */
-    public function __construct($targetDir)
+    public function __construct()
+    {
+        $this->targetDir = "";
+    }
+
+    /**
+     * @return string
+     */
+    public function getTargetDir()
+    {
+        return $this->targetDir;
+    }
+
+    /**
+     * @param string $targetDir
+     */
+    public function setTargetDir($targetDir)
     {
         $this->targetDir = $targetDir;
     }
@@ -30,23 +45,21 @@ class FileUploader
      */
     public function upload(UploadedFile $file)
     {
-        $fileName = md5(uniqid()).'.'.$file->guessExtension();
-
-        $file->move($this->getTargetDir(), $fileName);
-        return $fileName;
+        if(!empty($this->targetDir)) {
+            $fileName = md5(uniqid()).'.'.$file->guessExtension();
+            $file->move($this->getTargetDir(), $fileName);
+            return $fileName;
+        }
+        return false;
     }
 
     public function removeFile($fileName)
     {
-        $file_path = $this->getTargetDir().'/'.$fileName;
-        if(file_exists($file_path)) unlink($file_path);
-    }
-
-    /**
-     * @return mixed
-     */
-    public function getTargetDir()
-    {
-        return $this->targetDir;
+        if(!empty($this->targetDir)) {
+            $file_path = $this->getTargetDir().'/'.$fileName;
+            if(file_exists($file_path)) unlink($file_path);
+            return true;
+        }
+        return false;
     }
 }
